+++ src/manifest.js
