+++ src/merge.js
